Update givenergy_aio_ac3_balancer.yaml
Ensures that the Balancer Running toggle is off when Home Assistant boots. This is to prevent the scenario in which Home Assistant might have been restarted whilst balancing is running. If the toggle is not turned off on boot then the automation will fail to run and one of the batteries will never be used.